move pp to netman.pp
